Refactor atfork logic to elide locking on fast path
The atfork infrastructure is modified so that on Windows it
optimizes practically to nothing.

Otherwise the forkid is kept in a global variable, which is
updated exactly once per check round (to minimize syscalls).

A fast path to avoid locking is added. When the atfork check
is done, it is guaranteed that the P11_forkid does not change
(it can change only for the after-fork child process).

Only if a fork is detected, the locks are taken. Each object's
forkid is again checked in each of the check_*_int calls, so they
get updated only once even if the child is multithreaded.

This already improves performance measurably, but is even more
important to avoid lock contention after session pooling support
is added.